# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
  models.CarsCount:
    properties:
      q1:
        type: integer
      q2:
        type: integer
      q3:
        type: integer
      q4:
        type: integer
    type: object
...
  models.Inspection:
    properties:
      expiry_date:
        type: string
      inspection_date:
        type: string
      inspection_id:
        type: integer
      registration_id:
        type: string
      station_code:
        type: string
    type: object
  models.InspectionsList:
    properties:
      has_more:
        type: boolean
      inspections:
        items:
          $ref: '#/definitions/models.Inspection'
        type: array
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.QuarterAndYear:
    properties:
      cars_count:
        $ref: '#/definitions/models.CarsCount'
      year:
        type: integer
    type: object
...
  models.StationQuarterAndYear:
    properties:
      cars_count:
        $ref: '#/definitions/models.CarsCount'
      station_code:
        type: string
      year:
        type: integer
    type: object
...
info:
  contact:
    email: ddang.daodang@gmail.com
    name: Dang Dao
  description: RegistryTotal Golang REST API
  title: RegistryTotal REST API
  version: "1.0"
paths:
...
  /insp/:
    post:
      consumes:
      - application/json
      description: Create inspection
      operationId: Create inspection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Inspection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Create inspection
      tags:
      - insp
  /insp/{inspection_id}:
    delete:
      consumes:
      - application/json
      description: Delete inspection
      operationId: Delete
      parameters:
      - description: inspection id
        format: inspection_id
        in: path
        name: inspection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Delete inspection
      tags:
      - insp
    get:
      consumes:
      - application/json
      description: Get inspection by ID
      operationId: GetByID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InspectionsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get inspection by ID
      tags:
      - insp
    put:
      consumes:
      - application/json
      description: Update inspection
      operationId: Update
      parameters:
      - description: inspection id
        format: inspection_id
        in: path
        name: inspection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Inspection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Update inspection
      tags:
      - insp
  /insp/all:
    get:
      consumes:
      - application/json
      description: Get the list of all inspections
      operationId: GetAll
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InspectionsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get all inspections following pagination query(size, page, orderBy
      tags:
      - insp
  /insp/expiry-date:
    get:
      consumes:
      - application/json
      description: Get inspections by expiry date
      parameters:
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Inspection'
      summary: Get inspections by expiry date
      tags:
      - insp
  /insp/inspection-date:
    get:
      consumes:
      - application/json
      description: Get inspections by inspection date
      parameters:
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InspectionsList'
      summary: Get inspections by inspection date
      tags:
      - insp
  /insp/registration/{registration_id}:
    get:
      consumes:
      - application/json
      description: Get inspection by registration id
      operationId: GetByRegistrationID
      parameters:
      - description: registration id
        format: registration_id
        in: path
        name: registration_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InspectionsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get inspection by registration id
      tags:
      - insp
  /insp/station/{station_code}:
    get:
      consumes:
      - application/json
      description: Get inspection by station code
      operationId: GetByStationCode
      parameters:
      - description: area code
        format: station_code
        in: path
        name: station_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InspectionsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get inspection by station code
      tags:
      - insp
  /insp/statistic:
    get:
      consumes:
      - application/json
      description: Count inspections by quarter and year
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QuarterAndYear'
      summary: Count inspections by quarter and year
      tags:
      - insp
...
  /insp/statistic/station/{station_code}:
    get:
      description: Count inspections by quarter and year in station
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StationQuarterAndYear'
      summary: Count inspections by quarter and year in station
      tags:
      - insp
...
swagger: "2.0"